Circles and Info work.
Pompey and Ceasar done. Needed to make a global var "circleNum" If you
have any ideas about how to get it out let me know. In addition I
changed the HTML so that we only need one modal window. I pass an
additional arg to cellCliked with the ID of the corresponding circle to
enable or disable.